backend/requirements.txt
fastapi>=0.115.12
uvicorn>=0.27.1
playwright>=1.42.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
pydantic>=2.6.3
beautifulsoup4>=4.12.3
requests>=2.31.0
python-multipart>=0.0.9
aiohttp>=3.9.3
pillow>=10.2.0
google-cloud-aiplatform>=1.38.1 